+++ src/constants.strings.ts
@@ -1,0 +1,8 @@
+// Application string constants (or long-lived values)
+export const UNINSTANTIATED = `HumbleSDK is not instantiated!`;
+export const ASSURANCE_MSG = `Your funds were not moved.`;
+export const MIN_BALANCE_MSG = `Transaction would drop account below minimum required balance.`;
+export const POPUP_BLOCKED_MSG = `Your browser is blocking popups; please disable this blocker before retrying.`;
+export const TRANSACTION_CANCELLED_MSG = `Transaction Cancelled.`;
+export const TRANSACTION_DIDNT_LOAD = `Sorry, the transaction window didn't load due to some connection issues. Please try again.`;
+export const POOL_CREATION_ERR = `Pool creation failed.`;