+++ src/participants/index.ts
@@ -3,6 +3,7 @@
 export { fetchLiquidityPool, fetchToken } from "./PoolAnnouncer";
 export { createLiquidityPool } from "./PoolAdmin";
 export { createStakingPool } from "./StakingAdmin";
+export { CreateLimitOrderOpts, createLimitOrder } from "./LimitOrder";
 export {
   calculateRewardsPerBlock,
   checkRewardsImbalance